fix a bug on RegenerateRecentLogin

+++ app/Listeners/RegenerateRecentLogin.php
@@ -4,8 +4,9 @@ namespace App\Listeners;
 
 use App\Models\RecentLogin;
 use Illuminate\Auth\Events\Login;
+use Illuminate\Support\Facades\Auth;
 
-class AddUserToRecentLogins
+class RegenerateRecentLogin
 {
     /**
      * Create the event listener.
@@ -25,6 +26,14 @@ class AddUserToRecentLogins
      */
     public function handle(Login $event)
     {
+        //if user session has ended, next request will try to re-login via remember_me cookie
+        //and that will trigger the login event and this listener will be called
+        //since fingerprint and useragent can't be recorded into the recent login, it will throw error
+        //so we disabled re-generating recent login when user authenticated via remember
+        if (Auth::viaRemember()) {
+            return;
+        }
+
         $this->deleteRecentLogins($event->user);
 
         $recent = RecentLogin::createFromRequest([
